refactor: search input
